Add files via upload
The different files used for all calculations and figures in my article. 
Funcs are a number of diffrent functions called from the other. Raster scans does the parameter scans and statiscal analysis. f_of_numberOsc and nOsc2 does the analysis on how fast the algorithm converges, and intro_plot and fig_plot makes the rest of the figures. Julia_code is my first rewrite of the code base from python to julia.